+++ pyebsdindex/EBSDImage/IPFcolor.py
@@ -28,16 +28,13 @@ This license available with a FAQ at: https://openfontlicense.org
 SIL OPEN FONT LICENSE Version 1.1 - 26 February 2007
 '''
 
-from PIL import  Image, ImageDraw, ImageFont
-import os
 import matplotlib.colors as pltcolors
 import matplotlib.pyplot as plt
-from matplotlib.font_manager import findfont, FontProperties
-FONT = findfont(FontProperties(family='sans-serif', weight='bold'), fontext='ttf', )
 
 import numpy as np
 
 from pyebsdindex import rotlib
+from pyebsdindex.EBSDImage import scalarimage
 
 
 def makeipf(ebsddata, indexer, vector=np.array([0,0,1.0]), xsize = None, ysize = None, addscalebar=False):
@@ -82,53 +79,10 @@ def makeipf(ebsddata, indexer, vector=np.array([0,0,1.0]), xsize = None, ysize =
   ipf_out[0:npts*3] = ipfout[0:npts,:].flatten()
   ipf_out = ipf_out.reshape(ysize, xsize, 3)
   if addscalebar == True:
-    ipf_out = add_scalebar(ipf_out, indexer.fID.xStep, rescale=False)
+    ipf_out = scalarimage.add_scalebar(ipf_out, indexer.fID.xStep, rescale=False)
   return ipf_out
 
-def scalarimage(ebsddata, indexer, datafield='pq', xsize = None, ysize = None, addscalebar=False, cmap='viridis', datafieldindex=0):
-  npoints = ebsddata.shape[-1]
-  ipfout = ebsddata[-1][datafield]
-  if len(ipfout.shape) > 1:
-    ipfout = ipfout[:,datafieldindex]
 
-  ipfout = ipfout.astype(np.float32)
-  if datafield == 'fit':
-    mn = ipfout[ipfout < 179].mean()
-    std = ipfout[ipfout < 179].std()
-    norm = plt.Normalize(vmin=max(0.0, mn-3*std), vmax=mn+3*std)
-  else:
-    norm = plt.Normalize()
-  ipfout = np.array(norm(ipfout))
-  #ipfout -= ipfout.min()
-  #ipfout *= 1.0/ipfout.max()
-
-  cm = plt.colormaps[cmap]
-  ipfout = cm(ipfout)
-
-  if xsize is not None:
-    xsize = int(xsize)
-    # if ysize is None:
-    # print(ysize)
-  else:
-    xsize = indexer.fID.nCols
-    # xsize = int(npoints)
-    # ysize = 1
-
-  if ysize is not None:
-    ysize = int(ysize)
-  else:
-    ysize = int(npoints // xsize + np.int64((npoints % xsize) > 0))
-
-  ipf_out = np.zeros((ysize, xsize, 3), dtype=np.float32)
-  ipf_out = ipf_out.flatten()
-  npts = min(int(npoints), int(xsize * ysize))
-  # if int(xsize*ysize) < npoints:
-  #   npts = int(xsize*ysize)
-  ipf_out[0:npts * 3] = ipfout[0:npts, 0:3].flatten()
-  ipf_out = ipf_out.reshape(ysize, xsize, 3)
-  if addscalebar == True:
-    ipf_out = add_scalebar(ipf_out, indexer.fID.xStep, rescale=False)
-  return ipf_out
 
 
 def qu2ipf_cubic(quats, vector=np.array([0,0,1.0])):
@@ -416,85 +370,7 @@ def ipf_ledgend_hex(size=512):
   fig.savefig("IPFHex.pdf",bbox_inches=0, transparent=True)
   plt.close(1001)
 
-def add_scalebar(image, stepsize, rescale=True):
-  # image: grayscale or color image to add scale bar to.
-  # stepsize: size of a pixel in microns.
-  imshape = image.shape
-  channels = 1
-  if len(imshape) > 2:
-    channels = imshape[-1]
 
-  scale_bar_size, scale_bar_width_px, units = _round_scalebar(imshape[1], stepsize, imfract=0.33)
-  #print(scale_bar_size, scale_bar_width_px, units)
-  #scale_bar_height_px = np.int64(scale_bar_width_px / (16.18/2) ) # use golden ratio.
-  #underbar_size = (scale_bar_height_px*3, imshape[1])
-  underbar_size = (int(imshape[1] * 0.04) , imshape[1])
-  scale_bar_height_px = int(underbar_size[0] * 0.5)
-
-
-  underbar = np.zeros(underbar_size, dtype=np.uint8)+255
-  #yxoffset = (int(0.25*underbar_size[0]), int(0.01*imshape[1]))
-  yxoffset = (int(0.25 * underbar_size[0]), 0)
-  # add our scale bar.
-  underbar[yxoffset[0]:yxoffset[0]+scale_bar_height_px,
-              yxoffset[1]:yxoffset[1]+scale_bar_width_px]  = 0
-
-  underbarim = Image.fromarray(underbar, mode='L')
-  draw = ImageDraw.Draw(underbarim)
-  fontsize = scale_bar_height_px * 1.4
-
-  imfont = ImageFont.truetype(os.path.join(os.path.dirname(__file__), 'OpenSans-Bold.ttf'), fontsize)
-  #imfont = ImageFont.truetype(FONT, fontsize)
-  imtext = ' ' + str(scale_bar_size) + ' ' + units
-  text_color = 0
-  text_length = draw.textlength(imtext, imfont)
-  txoffset = (yxoffset[1]+scale_bar_width_px, yxoffset[0]-fontsize*0.35)
-  draw.text(txoffset, imtext, fill=text_color, font=imfont)
-
-  underbar = np.array(underbarim, dtype=np.float32)
-  underbar -= underbar.min()
-  underbar *= 1.0/underbar.max()
-
-
-
-  newshp = (imshape[0]+underbar_size[0], imshape[1], channels)
-  scalebarim = np.zeros(newshp, dtype=np.float32)
-
-  rescaleim = image.astype(np.float32)
-  if rescale == True:
-    rescaleim -= rescaleim.min()
-    rescaleim *= (1.0 / (rescaleim.max()))*0.999
-
-
-  rescaleim = rescaleim.reshape((imshape[0], imshape[1], channels))
-  scalebarim[0:imshape[0], 0:imshape[1], :] = rescaleim
-  for i in range(channels):
-    scalebarim[imshape[0]:, 0:imshape[1], i] = underbar
-  scalebarim = np.squeeze(scalebarim)
-  return scalebarim
-
-
-
-
-def _round_scalebar(image_width, pixel_width, imfract=0.333):
-    # image width is the width of the image in pixels (steps) in microns.
-    # pixel_width is the width of one pixel (aka step size).
-    # imfract is the maximum fraction of the image width that is desired for the scale bar size.
-    units = 'μm'
-    # these are the acceptable scale bar sizes
-    sequence = np.array([1,2,5,10,20,25,50,100,200,250,500], dtype=np.int64)
-    max_scale_bar_size = image_width*pixel_width*imfract
-    if max_scale_bar_size < 1:
-      units = 'nm'
-      max_scale_bar_size *= 1000.0
-      pixel_width *= 1000.0
-    if max_scale_bar_size > 1000:
-      units = 'mm'
-      max_scale_bar_size *= 0.001
-      pixel_width *= 0.001
-    scale_bar_size = sequence[sequence < max_scale_bar_size].max()
-    scale_bar_size_px = np.int64(np.float32(scale_bar_size/np.float32(pixel_width)))
-    return scale_bar_size, scale_bar_size_px, units
